+++ Salvo/Salvo/Repositories/IGamePlayerRepository.cs
@@ -8,7 +8,7 @@ namespace Salvo.Repositories
 {
     public interface IGamePlayerRepository
     {
-        public GamePlayer GetGamePlayerView(long idGamePlayer);
+        public GamePlayer GetGamePlayerView(int idGamePlayer);
 
         void Save(GamePlayer gamePlayer);
